[refactor] APIResponse 단일화
