+++ src/main/java/life/inha/icemarket/controller/ItemController.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import life.inha.icemarket.config.swagger.ApiDocumentResponse;
 import life.inha.icemarket.domain.Item;
 import life.inha.icemarket.dto.ItemSaveRequestDto;
 import life.inha.icemarket.service.ItemService;
@@ -24,11 +25,7 @@ public class ItemController {
     private final ItemService itemService;
 
     @Operation(description = "복지물품 추가")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "잘못된 유저 복지물품 추가 요청"),
-            @ApiResponse(responseCode = "403", description = "클라이언트의 접근 권한이 없음")
-    })
+    @ApiDocumentResponse
     @PostMapping("/create")
     public ResponseEntity<Item> createItem(@RequestBody ItemSaveRequestDto requestDto) {
         HttpHeaders header = new HttpHeaders();
@@ -45,11 +42,7 @@ public class ItemController {
     }
 
     @Operation(description = "복지물품 수정")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "잘못된 유저 복지물품 수정 요청"),
-            @ApiResponse(responseCode = "403", description = "클라이언트의 접근 권한이 없음")
-    })
+    @ApiDocumentResponse
     @PostMapping("/update/{id}")
     public ResponseEntity<Item> updateItem(@RequestBody ItemSaveRequestDto requestDto, @PathVariable Integer id) {
         HttpHeaders header = new HttpHeaders();
@@ -58,11 +51,7 @@ public class ItemController {
     }
 
     @Operation(description = "복지물품 삭제")
-    @ApiResponses({
-            @ApiResponse(responseCode = "200"),
-            @ApiResponse(responseCode = "400", description = "잘못된 유저 복지물품 삭제 요청"),
-            @ApiResponse(responseCode = "403", description = "클라이언트의 접근 권한이 없음")
-    })
+    @ApiDocumentResponse
     @GetMapping("/delete/{id}")
     public void deleteItem(@PathVariable Integer id) {
         itemService.delete(id);
